# backend/models/articles.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from backend.models.base_model_ import Model
from backend.models.article import Article  # noqa: F401,E501
from backend import util


class Articles(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, articles: List[Article]=None):  # noqa: E501
        """Articles - a model defined in Swagger

        :param articles: The articles of this Articles.  # noqa: E501
        :type articles: List[Article]
        """
        self.swagger_types = {
            'articles': List[Article]
        }

        self.attribute_map = {
            'articles': 'articles'
        }
        self._articles = articles

    @classmethod
    def from_dict(cls, dikt) -> 'Articles':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Articles of this Articles.  # noqa: E501
        :rtype: Articles
        """
        return util.deserialize_model(dikt, cls)

    @property
    def articles(self) -> List[Article]:
        """Gets the articles of this Articles.


        :return: The articles of this Articles.
        :rtype: List[Article]
        """
        return self._articles

    @articles.setter
    def articles(self, articles: List[Article]):
        """Sets the articles of this Articles.


        :param articles: The articles of this Articles.
        :type articles: List[Article]
        """

        self._articles = articles