# backend/models/auth.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from backend.models.base_model_ import Model
from backend import util


class Auth(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, login: str=None, password: str=None, captcha: str=None):  # noqa: E501
        """Auth - a model defined in Swagger

        :param login: The login of this Auth.  # noqa: E501
        :type login: str
        :param password: The password of this Auth.  # noqa: E501
        :type password: str
        :param captcha: The captcha of this Auth.  # noqa: E501
        :type captcha: str
        """
        self.swagger_types = {
            'login': str,
            'password': str,
            'captcha': str
        }

        self.attribute_map = {
            'login': 'login',
            'password': 'password',
            'captcha': 'captcha'
        }
        self._login = login
        self._password = password
        self._captcha = captcha

    @classmethod
    def from_dict(cls, dikt) -> 'Auth':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Auth of this Auth.  # noqa: E501
        :rtype: Auth
        """
        return util.deserialize_model(dikt, cls)

    @property
    def login(self) -> str:
        """Gets the login of this Auth.


        :return: The login of this Auth.
        :rtype: str
        """
        return self._login

    @login.setter
    def login(self, login: str):
        """Sets the login of this Auth.


        :param login: The login of this Auth.
        :type login: str
        """
        if login is None:
            raise ValueError("Invalid value for `login`, must not be `None`")  # noqa: E501

        self._login = login

    @property
    def password(self) -> str:
        """Gets the password of this Auth.


        :return: The password of this Auth.
        :rtype: str
        """
        return self._password

    @password.setter
    def password(self, password: str):
        """Sets the password of this Auth.


        :param password: The password of this Auth.
        :type password: str
        """
        if password is None:
            raise ValueError("Invalid value for `password`, must not be `None`")  # noqa: E501

        self._password = password

    @property
    def captcha(self) -> str:
        """Gets the captcha of this Auth.


        :return: The captcha of this Auth.
        :rtype: str
        """
        return self._captcha

    @captcha.setter
    def captcha(self, captcha: str):
        """Sets the captcha of this Auth.


        :param captcha: The captcha of this Auth.
        :type captcha: str
        """
        if captcha is None:
            raise ValueError("Invalid value for `captcha`, must not be `None`")  # noqa: E501

        self._captcha = captcha