# backend/models/domain_list.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from backend.models.base_model_ import Model
from backend.models.domains import Domains  # noqa: F401,E501
from backend.models.error_result import ErrorResult  # noqa: F401,E501
from backend import util


class DomainList(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, domains: List[str]=None, result: str=None, error: str=None):  # noqa: E501
        """DomainList - a model defined in Swagger

        :param domains: The domains of this DomainList.  # noqa: E501
        :type domains: List[str]
        :param result: The result of this DomainList.  # noqa: E501
        :type result: str
        :param error: The error of this DomainList.  # noqa: E501
        :type error: str
        """
        self.swagger_types = {
            'domains': List[str],
            'result': str,
            'error': str
        }

        self.attribute_map = {
            'domains': 'domains',
            'result': 'result',
            'error': 'error'
        }
        self._domains = domains
        self._result = result
        self._error = error

    @classmethod
    def from_dict(cls, dikt) -> 'DomainList':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The DomainList of this DomainList.  # noqa: E501
        :rtype: DomainList
        """
        return util.deserialize_model(dikt, cls)

    @property
    def domains(self) -> List[str]:
        """Gets the domains of this DomainList.


        :return: The domains of this DomainList.
        :rtype: List[str]
        """
        return self._domains

    @domains.setter
    def domains(self, domains: List[str]):
        """Sets the domains of this DomainList.


        :param domains: The domains of this DomainList.
        :type domains: List[str]
        """

        self._domains = domains

    @property
    def result(self) -> str:
        """Gets the result of this DomainList.


        :return: The result of this DomainList.
        :rtype: str
        """
        return self._result

    @result.setter
    def result(self, result: str):
        """Sets the result of this DomainList.


        :param result: The result of this DomainList.
        :type result: str
        """

        self._result = result

    @property
    def error(self) -> str:
        """Gets the error of this DomainList.


        :return: The error of this DomainList.
        :rtype: str
        """
        return self._error

    @error.setter
    def error(self, error: str):
        """Sets the error of this DomainList.


        :param error: The error of this DomainList.
        :type error: str
        """

        self._error = error