# backend/models/domains.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from backend.models.base_model_ import Model
from backend import util


class Domains(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, domains: List[str]=None):  # noqa: E501
        """Domains - a model defined in Swagger

        :param domains: The domains of this Domains.  # noqa: E501
        :type domains: List[str]
        """
        self.swagger_types = {
            'domains': List[str]
        }

        self.attribute_map = {
            'domains': 'domains'
        }
        self._domains = domains

    @classmethod
    def from_dict(cls, dikt) -> 'Domains':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Domains of this Domains.  # noqa: E501
        :rtype: Domains
        """
        return util.deserialize_model(dikt, cls)

    @property
    def domains(self) -> List[str]:
        """Gets the domains of this Domains.


        :return: The domains of this Domains.
        :rtype: List[str]
        """
        return self._domains

    @domains.setter
    def domains(self, domains: List[str]):
        """Sets the domains of this Domains.


        :param domains: The domains of this Domains.
        :type domains: List[str]
        """

        self._domains = domains