# backend/models/mam_message_list.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from backend.models.base_model_ import Model
from backend.models.mam_message import MAMMessage  # noqa: F401,E501
from backend import util


class MAMMessageList(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, messages: List[MAMMessage]=None):  # noqa: E501
        """MAMMessageList - a model defined in Swagger

        :param messages: The messages of this MAMMessageList.  # noqa: E501
        :type messages: List[MAMMessage]
        """
        self.swagger_types = {
            'messages': List[MAMMessage]
        }

        self.attribute_map = {
            'messages': 'messages'
        }
        self._messages = messages

    @classmethod
    def from_dict(cls, dikt) -> 'MAMMessageList':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The MAMMessageList of this MAMMessageList.  # noqa: E501
        :rtype: MAMMessageList
        """
        return util.deserialize_model(dikt, cls)

    @property
    def messages(self) -> List[MAMMessage]:
        """Gets the messages of this MAMMessageList.


        :return: The messages of this MAMMessageList.
        :rtype: List[MAMMessage]
        """
        return self._messages

    @messages.setter
    def messages(self, messages: List[MAMMessage]):
        """Sets the messages of this MAMMessageList.


        :param messages: The messages of this MAMMessageList.
        :type messages: List[MAMMessage]
        """

        self._messages = messages