# backend/models/mam_message_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from backend.models.base_model_ import Model
from backend import util


class MAMMessageRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, jid: str=None, ts: int=None):  # noqa: E501
        """MAMMessageRequest - a model defined in Swagger

        :param jid: The jid of this MAMMessageRequest.  # noqa: E501
        :type jid: str
        :param ts: The ts of this MAMMessageRequest.  # noqa: E501
        :type ts: int
        """
        self.swagger_types = {
            'jid': str,
            'ts': int
        }

        self.attribute_map = {
            'jid': 'jid',
            'ts': 'ts'
        }
        self._jid = jid
        self._ts = ts

    @classmethod
    def from_dict(cls, dikt) -> 'MAMMessageRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The MAMMessageRequest of this MAMMessageRequest.  # noqa: E501
        :rtype: MAMMessageRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def jid(self) -> str:
        """Gets the jid of this MAMMessageRequest.


        :return: The jid of this MAMMessageRequest.
        :rtype: str
        """
        return self._jid

    @jid.setter
    def jid(self, jid: str):
        """Sets the jid of this MAMMessageRequest.


        :param jid: The jid of this MAMMessageRequest.
        :type jid: str
        """
        if jid is None:
            raise ValueError("Invalid value for `jid`, must not be `None`")  # noqa: E501

        self._jid = jid

    @property
    def ts(self) -> int:
        """Gets the ts of this MAMMessageRequest.


        :return: The ts of this MAMMessageRequest.
        :rtype: int
        """
        return self._ts

    @ts.setter
    def ts(self, ts: int):
        """Sets the ts of this MAMMessageRequest.


        :param ts: The ts of this MAMMessageRequest.
        :type ts: int
        """

        self._ts = ts